More dynamic grid layout
This layout configuration / rendering is more closely aligned with
how I want to use it than what is fully possible or maybe even a
better usage of dockview. However, I have a very specific grid/panel
layout that I'm trying to achieve.

It can still be helpful to others though when trying to figure
out how to use dockview.